Split the API source file into 4 parts
